+++ chksum/src/lib.rs
@@ -1,4 +1,6 @@
 pub use chksum_arch as arch;
+#[cfg(feature = "async")]
+pub use chksum_async as r#async;
 pub use chksum_config as config;
 pub use chksum_hash as hash;
 #[cfg(feature = "sync")]