+++ chksum/src/prelude.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "async")]
+pub use crate::r#async::AsyncChksum;
 pub use crate::config::Config;
 pub use crate::hash::{Hash, Reset};
 #[cfg(feature = "sync")]